LayoutTests/editing/deleting/merge-lists.html

<!DOCTYPE html>
<html>
<body>
<p id="description">Placing cursor at the end of first list and executing forward delete should merge the second list with the first one.</p>
<div contenteditable>
<ol><li>one</li><li>two</li><li id="li">three</li></ol>
<ol><li>four</li><li>five</li><li>six</li></ol>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
var li = document.getElementById("li");
var selection = window.getSelection();
selection.collapse(li, li.childNodes.length);
document.execCommand("forwardDelete");
Markup.description(document.getElementById('description').textContent);
Markup.dump(document.querySelector('div'));
</script>
</body>
</html>
