third_party/blink/web_tests/editing/deleting/merge-lists-expected.txt

Placing cursor at the end of first list and executing forward delete should merge the second list with the first one.
| "
"
| <ol>
|   <li>
|     "one"
|   <li>
|     "two"
|   <li>
|     id="li"
|     "three<#selection-caret>"
|   <li>
|     "four"
|   <li>
|     "five"
|   <li>
|     "six"
| "
"
